Revamping rekor e2e - part 4 of N
- Updated tests for ssh
- Updated tests for minisign
