+++ tests/e2e_test.go
@@ -37,7 +37,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
@@ -236,138 +235,6 @@ func TestGetCLI(t *testing.T) {
 
 func TestSearchNoEntriesRC1(t *testing.T) {
 	runCliErr(t, "search", "--email", "noone@internetz.com")
-}
-
-func TestMinisign(t *testing.T) {
-	// Create a keypair
-	keyPath := filepath.Join(t.TempDir(), "minisign.key")
-	pubPath := filepath.Join(t.TempDir(), "minisign.pub")
-
-	// Set an empty password, we have to hit enter twice to confirm
-	run(t, "\n\n", "minisign", "-G", "-s", keyPath, "-p", pubPath)
-
-	// Create a random artifact and sign it.
-	artifactPath := filepath.Join(t.TempDir(), "artifact")
-	sigPath := filepath.Join(t.TempDir(), "signature.asc")
-	createArtifact(t, artifactPath)
-
-	// Send in one empty password over stdin
-	out := run(t, "\n", "minisign", "-S", "-s", keyPath, "-m", artifactPath, "-x", sigPath)
-	t.Log(out)
-
-	// Now upload to the log!
-	out = runCli(t, "upload", "--artifact", artifactPath, "--signature", sigPath,
-		"--public-key", pubPath, "--pki-format", "minisign")
-	outputContains(t, out, "Created entry at")
-
-	uuidA := getUUIDFromUploadOutput(t, out)
-
-	out = runCli(t, "verify", "--artifact", artifactPath, "--signature", sigPath,
-		"--public-key", pubPath, "--pki-format", "minisign")
-	outputContains(t, out, "Inclusion Proof")
-
-	out = runCli(t, "search", "--public-key", pubPath, "--pki-format", "minisign")
-	outputContains(t, out, uuidA)
-
-	// crease a second artifact and sign it
-	artifactPath_B := filepath.Join(t.TempDir(), "artifact2")
-	createArtifact(t, artifactPath_B)
-	out = run(t, "\n", "minisign", "-S", "-s", keyPath, "-m", artifactPath_B, "-x", sigPath)
-	// Now upload to the log!
-	out = runCli(t, "upload", "--artifact", artifactPath_B, "--signature", sigPath,
-		"--public-key", pubPath, "--pki-format", "minisign")
-	outputContains(t, out, "Created entry at")
-	uuidB := getUUIDFromUploadOutput(t, out)
-
-	tests := []struct {
-		name               string
-		expectedUuidACount int
-		expectedUuidBCount int
-		artifact           string
-		operator           string
-	}{
-		{
-			name:               "artifact A AND signature should return artifact A",
-			expectedUuidACount: 1,
-			expectedUuidBCount: 0,
-			artifact:           artifactPath,
-			operator:           "and",
-		},
-		{
-			name:               "artifact A OR signature should return artifact A and B",
-			expectedUuidACount: 1,
-			expectedUuidBCount: 1,
-			artifact:           artifactPath,
-			operator:           "or",
-		},
-		{
-			name:               "artifact B AND signature should return artifact B",
-			expectedUuidACount: 0,
-			expectedUuidBCount: 1,
-			artifact:           artifactPath_B,
-			operator:           "and",
-		},
-		{
-			name:               "artifact B OR signature should return artifact A and B",
-			expectedUuidACount: 1,
-			expectedUuidBCount: 1,
-			artifact:           artifactPath_B,
-			operator:           "or",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			out = runCli(t, "search", "--public-key", pubPath, "--pki-format", "minisign",
-				"--operator", test.operator, "--artifact", test.artifact)
-
-			expected := map[string]int{uuidA: test.expectedUuidACount, uuidB: test.expectedUuidBCount}
-			actual := map[string]int{
-				uuidA: strings.Count(out, uuidA),
-				uuidB: strings.Count(out, uuidB),
-			}
-			if !reflect.DeepEqual(expected, actual) {
-				t.Errorf("expected to find %v, found %v", expected, actual)
-			}
-		})
-	}
-}
-
-func TestSSH(t *testing.T) {
-	td := t.TempDir()
-	// Create a keypair
-	keyPath := filepath.Join(td, "id_rsa")
-	pubPath := filepath.Join(td, "id_rsa.pub")
-
-	if err := ioutil.WriteFile(pubPath, []byte(sshPublicKey), 0600); err != nil {
-		t.Fatal(err)
-	}
-	if err := ioutil.WriteFile(keyPath, []byte(sshPrivateKey), 0600); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a random artifact and sign it.
-	artifactPath := filepath.Join(td, "artifact")
-	sigPath := filepath.Join(td, "signature.sig")
-	artifact := createArtifact(t, artifactPath)
-
-	sig := SSHSign(t, strings.NewReader(artifact))
-	if err := ioutil.WriteFile(sigPath, []byte(sig), 0600); err != nil {
-		t.Fatal(err)
-	}
-
-	// Now upload to the log!
-	out := runCli(t, "upload", "--artifact", artifactPath, "--signature", sigPath,
-		"--public-key", pubPath, "--pki-format", "ssh")
-	outputContains(t, out, "Created entry at")
-
-	uuid := getUUIDFromUploadOutput(t, out)
-
-	out = runCli(t, "verify", "--artifact", artifactPath, "--signature", sigPath,
-		"--public-key", pubPath, "--pki-format", "ssh")
-	outputContains(t, out, "Inclusion Proof")
-
-	out = runCli(t, "search", "--public-key", pubPath, "--pki-format", "ssh")
-	outputContains(t, out, uuid)
 }
 
 func TestIntoto(t *testing.T) {
